inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
transformers==4.46.0
torch==2.5.0
Pillow==10.4.0
numpy==1.26.4
huggingface_hub==0.25.0
pytest==8.3.0
httpx==0.27.0
